ansible/docker/requirements.txt
gitpython==2.1.11
docker==3.7.2
